move shared avatar properties/methods to base class
